refactor: UserDao 공통로직 JdbcTemplate 으로 이전
