+++ app/src/main/java/com/techcourse/dao/UserDao.java
@@ -1,11 +1,6 @@
 package com.techcourse.dao;
 
 import com.techcourse.domain.User;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
 import javax.sql.DataSource;
 import nextstep.jdbc.JdbcTemplate;
@@ -16,11 +11,9 @@ public class UserDao {
 
     private static final Logger log = LoggerFactory.getLogger(UserDao.class);
 
-    private final DataSource dataSource;
     private final JdbcTemplate jdbcTemplate;
 
     public UserDao(DataSource dataSource) {
-        this.dataSource = dataSource;
         this.jdbcTemplate = new JdbcTemplate(dataSource);
     }
 
@@ -39,92 +32,39 @@ public class UserDao {
     public List<User> findAll() {
         final String sql = "select * from users";
 
-        try (Connection conn = dataSource.getConnection();
-                PreparedStatement pstmt = conn.prepareStatement(sql);
-                ResultSet rs = pstmt.executeQuery()) {
-
-            log.debug("query : {}", sql);
-
-            List<User> users = new ArrayList<>();
-            while (rs.next()) {
-                long id = rs.getLong("id");
-                String account = rs.getString("account");
-                String password = rs.getString("password");
-                String email = rs.getString("email");
-                User user = new User(id, account, password, email);
-                users.add(user);
-            }
-            return users;
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new RuntimeException(e);
-        }
+        return jdbcTemplate.queryForList(sql,
+                (rs) -> new User(
+                        rs.getLong(1),
+                        rs.getString(2),
+                        rs.getString(3),
+                        rs.getString(4)));
     }
 
     public User findById(Long id) {
         final String sql = "select id, account, password, email from users where id = ?";
 
-        try (Connection conn = dataSource.getConnection();
-                PreparedStatement pstmt = createPreparedStatement(conn, sql, id);
-                ResultSet rs = pstmt.executeQuery()
-        ) {
-            log.debug("query : {}", sql);
-
-            if (rs.next()) {
-                return new User(
+        return jdbcTemplate.query(sql,
+                (rs) -> new User(
                         rs.getLong(1),
                         rs.getString(2),
                         rs.getString(3),
-                        rs.getString(4));
-            }
-            return null;
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new RuntimeException(e);
-        }
+                        rs.getString(4)), id);
     }
 
     public User findByAccount(String account) {
         final String sql = "select id, account, password, email from users where account = ?";
 
-        try (Connection conn = dataSource.getConnection();
-                PreparedStatement pstmt = createPreparedStatement(conn, sql, account);
-                ResultSet rs = pstmt.executeQuery()) {
-
-            log.debug("query : {}", sql);
-
-            if (rs.next()) {
-                return new User(
+        return jdbcTemplate.query(sql,
+                (rs) -> new User(
                         rs.getLong(1),
                         rs.getString(2),
                         rs.getString(3),
-                        rs.getString(4));
-            }
-            return null;
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new RuntimeException(e);
-        }
+                        rs.getString(4)), account);
     }
 
     public void removeAll() {
         final String sql = "delete from users";
 
         jdbcTemplate.update(sql);
-    }
-
-    private PreparedStatement createPreparedStatement(Connection conn, String sql, Object... args) throws SQLException {
-        PreparedStatement pstmt = conn.prepareStatement(sql);
-        int index = 1;
-        for (Object arg : args) {
-            if (arg instanceof Long) {
-                pstmt.setLong(index, (Long) arg);
-            }
-            if (arg instanceof String) {
-                pstmt.setString(index, (String) arg);
-            }
-            index += 1;
-        }
-        return pstmt;
     }
 }
