+++ jdbc/src/main/java/nextstep/jdbc/JdbcTemplate.java
@@ -2,7 +2,10 @@ package nextstep.jdbc;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
 import javax.sql.DataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -31,6 +34,43 @@ public class JdbcTemplate {
         }
     }
 
+    public <T> T query(String sql, RowMapper<T> rowMapper, Object... args) {
+        try (Connection conn = dataSource.getConnection();
+                PreparedStatement pstmt = createPreparedStatement(conn, sql, args);
+                ResultSet rs = pstmt.executeQuery()
+        ) {
+            log.debug("query : {}", sql);
+
+            if (rs.next()) {
+                return rowMapper.mapRow(rs);
+            }
+            return null;
+        } catch (SQLException e) {
+            log.error(e.getMessage(), e);
+            throw new RuntimeException(e);
+        }
+    }
+
+    public <T> List<T> queryForList(String sql, RowMapper<T> rowMapper) {
+        try (Connection conn = dataSource.getConnection();
+                PreparedStatement pstmt = conn.prepareStatement(sql);
+                ResultSet rs = pstmt.executeQuery()) {
+
+            log.debug("query : {}", sql);
+
+            List<T> users = new ArrayList<>();
+            while (rs.next()) {
+                T t = rowMapper.mapRow(rs);
+                users.add(t);
+            }
+            return users;
+        } catch (SQLException e) {
+            log.error(e.getMessage(), e);
+            throw new RuntimeException(e);
+        }
+    }
+
+
     private void setPreparedStatement(PreparedStatement pstmt, Object[] args) throws SQLException {
         int index = 1;
         for (Object arg : args) {
@@ -42,5 +82,21 @@ public class JdbcTemplate {
             }
             index += 1;
         }
+    }
+
+
+    private PreparedStatement createPreparedStatement(Connection conn, String sql, Object... args) throws SQLException {
+        PreparedStatement pstmt = conn.prepareStatement(sql);
+        int index = 1;
+        for (Object arg : args) {
+            if (arg instanceof Long) {
+                pstmt.setLong(index, (Long) arg);
+            }
+            if (arg instanceof String) {
+                pstmt.setString(index, (String) arg);
+            }
+            index += 1;
+        }
+        return pstmt;
     }
 }
